--- src/main/java/com/relpy/comment/dao/CommentDAO.java
@@ -1,9 +1,0 @@
-package com.relpy.comment.dao;
-
-import java.util.List;
-
-import com.relpy.comment.to.Comment;
-
-public interface CommentDAO {
-	public List<Comment> findByThreadId(int id);
-}
